docs: add theme switcher on mobile, fix tabs demo wrong import
