+++ packages/ui-core/src/components/Tabs/doc/Tabs.example.mdx
@@ -1,7 +1,7 @@
 import { Playground } from 'docz';
 import Tabs from '../Tabs';
 import Tab from '../Tab';
-import { DemoContent, OuterControlContentContainer, increaseSizeComponent, OuterContentContainer } from './Tabs.docz';
+import { DemoContent, OuterControlContentContainer, increaseSizeComponent, OuterObserveContainer } from './Tabs.docz';
 import Balance from '@megafon/ui-icons/basic-24-balance_24.svg';
 import ContentArea from 'components/ContentArea/ContentArea';
 import Button from 'components/Button/Button';
