+++ src/gatsby-theme-docz/components/SideBar/SideBar.less
@@ -14,6 +14,7 @@
 
     &__logo {
         padding: 16px 16px 16px 8px;
+        margin-right: 8px;
     }
 
     &__menu {
@@ -23,121 +24,6 @@
         @media screen and (max-width: 768px) {
             max-height: calc(100vh - 88px - 48px);
             margin-top: 24px;
-        }
-    }
-    &__theme-switcher-wrapper{
-        display: inline-block;
-        width: 56px;
-        height: 40px;
-        margin-left: 24px;
-        padding: 8px 10px 14px 8px;
-        cursor: pointer;
-    }
-    &__theme-switcher {
-        position: relative;
-        width: 34px;
-        height: 22px;
-        padding-top: 12px;
-        border: none;
-        border-radius: 11px;
-
-        background-color: var(--spbSky1);
-        cursor: pointer;
-
-        transition: background-color 0.3s;
-
-        @media screen and (max-width: 768px) {
-            margin-top: 0px;
-        }
-    }
-
-    &__theme-switcher:hover,
-    &__theme-switcher_dark {
-        background-color: var(--spbSky2);
-    }
-
-    &__theme-switcher_dark:hover {
-        background-color: var(--spbSky3);
-    }
-
-    &__theme-switcher:disabled {
-        background-color: var(--spbSky1);
-        cursor: auto;
-    }
-
-    &__switch-icons {
-        position: absolute;
-        top: 2px;
-        right: 2px;
-        z-index: 1;
-
-        width: 18px;
-        height: 18px;
-        border-radius: 50%;
-
-        background-color: var(--brandGreen);
-        transform: none;
-
-        transition: transform 0.4s cubic-bezier(0.25, 0.1, 0.25, 1), background-color 0.3s linear;
-    }
-
-    &__switch-icon {
-        position: absolute;
-        top: 50%;
-        left: 50%;
-
-        width: 13px;
-        height: 13px;
-
-        transform: translate(-50%, -50%);
-        opacity: 0;
-
-        transition: opacity 0.2s;
-    }
-
-    &__switch-dot {
-        position: absolute;
-        top: 50%;
-        right: 4px;
-        z-index: 0;
-
-        display: block;
-        width: 6px;
-        height: 6px;
-        border-radius: 50%;
-
-        background-color: var(--brandGreen);
-        transform: translateY(-50%);
-    }
-
-    &__switch-dot_blue {
-        right: auto;
-        left: 4px;
-
-        background-color: var(--reflexBlue);
-    }
-
-    &__theme-switcher_light & {
-        &__switch-icon_light {
-            opacity: 1;
-        }
-    }
-
-    &__theme-switcher_dark & {
-        &__switch-icon_dark {
-            opacity: 1;
-        }
-
-        &__switch-icons {
-            background-color: var(--reflexBlue);
-            transform: translateX(-12px);
-        }
-    }
-
-    &__theme-switcher:disabled & {
-        &__switch-icons,
-        &__switch-dot {
-            background-color: var(--spbSky2);
         }
     }
 }
